--- src/main/java/ru/aston/mcs/repository/ManagerRepository.java
@@ -1,9 +1,0 @@
-package ru.aston.mcs.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.aston.mcs.entity.Manager;
-
-@Repository
-public interface ManagerRepository extends JpaRepository<Manager, Long> {
-}
