--- src/main/java/ru/aston/mcs/service/ManagerService.java
@@ -1,19 +1,0 @@
-package ru.aston.mcs.service;
-
-
-import ru.aston.mcs.dto.ManagerDTO;
-
-import java.util.List;
-
-public interface ManagerService {
-
-    List<ManagerDTO> getAllManagers();
-
-    ManagerDTO getManager(Long managerId);
-
-    void addAndSaveManager(ManagerDTO manager);
-
-    void deleteManager(Long managerId);
-
-
-}
