+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=9005
+server.port=8080
 spring.datasource.url=${SPRING_DATASOURCE_URL}
 spring.datasource.username=${SPRING_DATASOURCE_USERNAME}
 spring.datasource.password=${SPRING_DATASOURCE_PASSWORD}
