+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,3 +1,3 @@
 databaseChangeLog:
   - include:
-      file: db/changelog/v1/create_shema/01-create-schema-changelog.yaml
+      file: db/changelog/v1/create_schema/01-create-schema-changelog.yaml
