+++ src/test/java/ru/aston/mcs/McsApplicationTests.java
@@ -1,4 +1,4 @@
-package ru.aston.mcs.service;
+package ru.aston.mcs;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -9,5 +9,4 @@ class McsApplicationTests {
     @Test
     void contextLoads() {
     }
-
 }
